finetune/pyproject.toml
[tool.poetry]
name = "testttt"
version = "0.1.0"
description = ""
authors = ["Preslav <preslav.aleksandrov@protonmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.11.7"
transformers = "^4.42.3"
flwr = {extras = ["simulation"], version = "^1.9.0" } 
flwr-datasets = {version = ">=0.0.2,<1.0.0" }
hydra-core = "^1.3.2"
trl = "^0.8.6"
bitsandbytes = "^0.43.0"
scipy = "^1.13.0"
peft = "^0.6.2"
gradio = "^3.50.2"
scikit-learn = "^1.4.0"
cryptography = "^42.0.7"
setuptools = "^70.0.0"
wandb = "^0.17.1"
opacus = "^1.4.1"
torch = "^2.3.1"
datasets = "2.20.0"
omegaconf = "^2.3.0"
huggingface-hub = "^0.23.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



